rename web_scraper.py to main.py
